+++ src/SetFolderPopup.cpp
@@ -1,16 +1,14 @@
 #include <Geode/Geode.hpp>
-#include <Geode/modify/SetIDPopup.hpp>
+#include <Geode/modify/SetFolderPopup.hpp>
 #include "Utils.hpp"
 
 using namespace geode::prelude;
 
-class $modify(SetIDPopup) {
+class $modify(SetFolderPopup) {
 
-	bool init(int current, int begin, int end, gd::string title, gd::string button, bool p5, int p6, float p7, bool p8, bool p9) {
-		if (!SetIDPopup::init(current, begin, end, title, button, p5, p6, p7, p8, p9))
+	bool init(int value, bool isCreated, gd::string title) {
+		if (!SetFolderPopup::init(value, isCreated, title))
 		return false;
-
-	// This also gets SetFolderPopup
 
 	auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
 	DayAndNightSystemOverlay->setZOrder(106);